Inline LegoAnimPresenters
