+++ LEGO1/lego/legoomni/src/actors/act3actors.cpp
@@ -4,9 +4,9 @@
 #include "act3ammo.h"
 #include "anim/legoanim.h"
 #include "define.h"
+#include "legoanimpresenter.h"
 #include "legobuildingmanager.h"
 #include "legocachesoundmanager.h"
-#include "legolocomotionanimpresenter.h"
 #include "legopathedgecontainer.h"
 #include "legoplantmanager.h"
 #include "legoplants.h"
