+++ LEGO1/lego/legoomni/src/worlds/legoact2.cpp
@@ -7,10 +7,10 @@
 #include "islepathactor.h"
 #include "jukebox_actions.h"
 #include "legoanimationmanager.h"
+#include "legoanimpresenter.h"
 #include "legocachesoundmanager.h"
 #include "legogamestate.h"
 #include "legoinputmanager.h"
-#include "legolocomotionanimpresenter.h"
 #include "legomain.h"
 #include "legopathstruct.h"
 #include "legosoundmanager.h"
